Simplify the definition of Portable{Collection,Object}<T, TDev> alias templates
Also move the CopyTo{Host,Device} specialisations to interface/ as
they are independent from ALPAKA_ACCELERATOR_NAMESPACE.
